Moved hands into their own gadget
This lets master just be master so it doesn't need to be shared. it also makes swapping out the default hands easier down the road.
