+++ websrc/default_hands/src/default_hands.css
@@ -1,0 +1,53 @@
+body, html
+{
+	background-color: transparent;
+	width: 100%;
+	height: 100%;
+	overflow: hidden;
+}
+
+.FullPage
+{
+	width: 100%;
+	height: 100%;
+	overflow: hidden;
+}
+
+.HookCircle
+{
+	border-radius: 50%;
+	border-width: 0.1rem;
+	border-style: solid;
+	width: 90%;
+	height: 90%;
+	font-size: 2rem;
+	margin: 1px;
+}
+
+.NoHighlight
+{
+	border-color: #88888888;
+	background-color: #88888855;
+}
+
+.InRangeHighlight
+{
+	border-color: #888888FF;
+	background-color: #888888AA;
+}
+
+.GadgetSeedContainer
+{
+	background-color: transparent;
+	display: flex;
+	flex-direction: row;
+	flex-wrap: wrap;
+	height: 100%;
+	width: 100%;
+}
+
+.GadgetSeed
+{
+	height: 30%;
+	width: 30%;
+}
