+++ websrc/server/persistence.ts
@@ -258,6 +258,7 @@ class CPersistenceManager
 	{
 		return this.m_state.installedGadgets.includes( gadgetUri )
 			|| gadgetUri == "http://localhost:23842/gadgets/aardvark_master"
+			|| gadgetUri == "http://localhost:23842/gadgets/default_hands"
 			|| gadgetUri == "http://localhost:23842/gadgets/gadget_menu";
 	}
