+++ websrc/webpack.config.js
@@ -94,6 +94,7 @@ module.exports =
 	createConfig( 'aardvark_master', 'Master App', 'tsx' ),
 	createConfig( 'aardvark_monitor', 'Monitor', 'tsx' ),
 	createConfig( 'aardvark_renderer', 'Renderer', 'ts' ),
+	createConfig( 'default_hands', 'Default Hands', 'tsx' ),
 	createConfig( 'test_panel', 'Test Panel', 'tsx' ),
 	createConfig( 'gadget_installer', 'Gadget Installer', 'tsx' ),
 	createConfig( 'gadget_menu', 'Gadget Menu', 'tsx' ),
